requirements.txt
pandas
plotly
streamlit
python-dotenv
openai
beautifulsoup4
lxml